requirements.txt
asttokens>=2.2.1
backcall>=0.2.0
certifi>=2022.12.7
charset-normalizer>=3.1.0
colorama>=0.4.6
comm>=0.1.2
debugpy>=1.6.6
decorator>=5.1.1
executing>=1.2.0
idna>=3.4
ipykernel>=6.21.3
ipython>=8.11.0
jedi>=0.18.2
jupyter_client>=8.0.3
jupyter_core>=5.2.0
matplotlib-inline>=0.1.6
nest-asyncio>=1.5.6
numpy>=1.24.2
packaging>=23.0
pandas>=1.5.3
parso>=0.8.3
pickleshare>=0.7.5
platformdirs>=3.1.0
praw>=7.7.0
prawcore>=2.3.0
prompt-toolkit>=3.0.38
psutil>=5.9.4
pure-eval>=0.2.2
Pygments>=2.14.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pytz>=2022.7.1
pywin32>=305
pyzmq>=25.0.0
requests>=2.28.2
six>=1.16.0
stack-data>=0.6.2
tornado>=6.2
traitlets>=5.9.0
update-checker>=0.18.0
urllib3>=1.26.14
wcwidth>=0.2.6
websocket-client>=1.5.1
